Test quick timing of GH cron workflows
Last time I dealt with it, GH cron workflows are documented to not schedule well below the every 10 minute threshold, and do not in practice schedule reliably below the every hour threshold.

I want to test this again, because I want to compare to how a runner managed with cirun.io performs vs. GH-hosted.